pyproject.toml
[project]
name = "homebase"
version = "0.1.0"
description = "CLI to aggregate assignments into Obsidian"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8",
  "jinja2>=3.1",
  "python-dotenv>=1.0",
  "tzdata>=2024.1; platform_system == 'Windows'"
]

[project.scripts]
homebase = "homebase.cli:app"

[tool.setuptools]
packages = ["homebase"]

[tool.setuptools.package-data]
# include templates as non-code data
"homebase" = ["../templates/*.j2"]
